Updating WinUI and creating MainPage
