+++ Simple Icon File Maker/Simple Icon File Maker (Package)/Simple Icon File Maker (Package).wapproj
@@ -134,7 +134,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.2.230217.4">
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.3.230331000">
       <IncludeAssets>build</IncludeAssets>
     </PackageReference>
     <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.755">
