+++ Simple Icon File Maker/Simple Icon File Maker/Simple Icon File Maker.csproj
@@ -21,6 +21,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Remove="AboutDialog.xaml" />
+    <None Remove="MainPage.xaml" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="SimpleIconMaker.ico" />
@@ -28,7 +29,7 @@
 
   <ItemGroup>
     <PackageReference Include="Magick.NET-Q16-AnyCPU" Version="12.3.0" />
-    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.2.230217.4" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.3.230331000" />
     <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.755" />
     <PackageReference Include="PropertyChanged.Fody" Version="4.1.0" />
     <Manifest Include="$(ApplicationManifest)" />
@@ -39,6 +40,12 @@
       <Pack>True</Pack>
       <PackagePath>\</PackagePath>
     </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Update="MainPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
 
   <ItemGroup>
